work on dao implementation of getting food items

+++ Rebecca_Morse_Code/CafeApp/src/main/java/com/cafe/service/MenuService.java
@@ -16,15 +16,15 @@ public class MenuService {
 
 		Map<Integer, FoodItem> menuContents = new HashMap<Integer, FoodItem>();
 
-		FoodItem water = new FoodItem("Water", "Beverage", "Liquid Refreshment at its' finest!", 0);
-		FoodItem lemonade = new FoodItem("Lemonade", "Beverage", "The tastiest Summertime Drink", 2);
+		FoodItem water = new FoodItem("Water", "Beverage", "Liquid Refreshment at its' finest!", (float) 0.00);
+		FoodItem lemonade = new FoodItem("Lemonade", "Beverage", "The tastiest Summertime Drink", (float) 2.00);
 		FoodItem veggieBurger = new FoodItem("Veggie Burger", "Entree",
-				"Our House-made Veggie Burgers, served with hand-cut fries.", 10);
-		FoodItem pasta = new FoodItem("Pasta Supreme", "Entree", "House made pasta with the freshest Pesto around.", 7);
-		FoodItem asparagus = new FoodItem("Asparagus Spear", "Side", "Only the freshest!", 17);
+				"Our House-made Veggie Burgers, served with hand-cut fries.", (float) 10.00);
+		FoodItem pasta = new FoodItem("Pasta Supreme", "Entree", "House made pasta with the freshest Pesto around.", (float) 7.00);
+		FoodItem asparagus = new FoodItem("Asparagus Spear", "Side", "Only the freshest!", (float) 17.00);
 		FoodItem cococadoShake = new FoodItem("Cococado Vegan Milkshake", "Dessert",
-				"Made with rice dream ice cream, bananas, coconut and avocado.", 4);
-		FoodItem pie = new FoodItem("Pie", "Dessert", "Fresh seasonal pie.  Current variety: peach", 5);
+				"Made with rice dream ice cream, bananas, coconut and avocado.", (float) 4.00);
+		FoodItem pie = new FoodItem("Pie", "Dessert", "Fresh seasonal pie.  Current variety: peach", (float) 5.00);
 
 		menuContents.put(1, water);
 		menuContents.put(2, lemonade);
